+++ product/agent/agent-model/src/main/java/io/codekvast/javaagent/model/Endpoints.java
@@ -33,6 +33,7 @@ public interface Endpoints {
         String V1_POLL_CONFIG = "/javaagent/v1/pollConfig";
         String V1_UPLOAD_CODEBASE = "/javaagent/v1/uploadCodeBase";
         String V2_UPLOAD_CODEBASE = "/javaagent/v2/uploadCodeBase";
+        String V3_UPLOAD_CODEBASE = "/javaagent/v3/uploadCodeBase";
         String V1_UPLOAD_INVOCATION_DATA = "/javaagent/v1/uploadInvocationData";
         String V2_UPLOAD_INVOCATION_DATA = "/javaagent/v2/uploadInvocationData";