+++ product/agent/java-agent/src/main/java/io/codekvast/javaagent/codebase/CodeBaseScanner.java
@@ -23,7 +23,7 @@ package io.codekvast.javaagent.codebase;
 
 import com.google.common.io.Files;
 import com.google.common.reflect.ClassPath;
-import io.codekvast.javaagent.model.v2.MethodSignature2;
+import io.codekvast.javaagent.model.v3.MethodSignature3;
 import io.codekvast.javaagent.util.SignatureUtils;
 import lombok.Builder;
 import lombok.Value;
@@ -246,7 +246,7 @@ public class CodeBaseScanner {
             Constructor[] declaredConstructors = clazz.getDeclaredConstructors();
 
             for (Constructor constructor : declaredConstructors) {
-                MethodSignature2 thisSignature = SignatureUtils.makeConstructorSignature(clazz, constructor);
+                MethodSignature3 thisSignature = SignatureUtils.makeConstructorSignature(clazz, constructor);
                 codeBase.addSignature(thisSignature);
             }
 
@@ -269,7 +269,7 @@ public class CodeBaseScanner {
             Method[] declaredMethods = clazz.getDeclaredMethods();
             for (Method method : declaredMethods) {
                 boolean ignored = true;
-                MethodSignature2 signature = SignatureUtils.makeMethodSignature(clazz, method);
+                MethodSignature3 signature = SignatureUtils.makeMethodSignature(clazz, method);
 
                 String declaringPackage = method.getDeclaringClass().getPackage().getName();
                 for (String pkg : packages) {