+++ product/agent/java-agent/src/main/java/io/codekvast/javaagent/config/AgentConfig.java
@@ -129,7 +129,7 @@ public class AgentConfig implements Serializable {
     }
 
     public String getCodeBaseUploadEndpoint() {
-        return String.format("%s%s", serverUrl, V2_UPLOAD_CODEBASE);
+        return String.format("%s%s", serverUrl, V3_UPLOAD_CODEBASE);
     }
 
     public String getInvocationDataUploadEndpoint() {